make EMT_ID a BIGINT rather than just INT.  Likely doesn't matter, but the fact that every Event Save results in deleted relations before the new relations get setup means that there are a lot of auto_increments happening in each save.  So ID could quickly increment when we eventually add more messengers! [touch: 5474]

+++ core/data_migration_scripts/EE_DMS_Core_4_4_0.dms.php
@@ -235,7 +235,7 @@ class EE_DMS_Core_4_4_0 extends EE_Data_Migration_Script_Base{
 		$this->_table_should_exist_previously( $table_name, $sql, 'ENGINE=InnoDB');
 
 		$table_name = 'esp_event_message_template';
-		$sql = "EMT_ID int(10) unsigned NOT NULL AUTO_INCREMENT,
+		$sql = "EMT_ID BIGINT(20) unsigned NOT NULL AUTO_INCREMENT,
 					EVT_ID BIGINT(20) unsigned NOT NULL DEFAULT 0,
 					GRP_ID INT(10) unsigned NOT NULL DEFAULT 0,
 					PRIMARY KEY (EMT_ID),
